added usage commnets to all functions

+++ circle.cpp
@@ -2,11 +2,23 @@
 #include <math.h>
 
 
-int Area(int r) {
+/**
+ * Calculates area of a circle
+ * 
+ * @param r Contains radius of a circle
+ * @return area of a circle with given radius
+ */
+float Area(int r) {
     return M_PI * r * r;
 }
 
 
-int Perimeter(int r) {
+/**
+ * Calculates perimeter of a circle
+ * 
+ * @param r Contains radius of a circle
+ * @return perimeter of a circle with given radius
+ */
+float Perimeter(int r) {
     return 2 * r * M_PI;
 }
